docs(Ascii): Add visualization docs for Architecture, DataStructure and AppDisplay

+++ src/core/components/PlayerAction.h
@@ -24,8 +24,8 @@ typedef struct
 {
     ActionType actionType; // P, L, D
     PlayerId playerId;     // White or black
-    int squareIdx;         // Square
-    int topStoneIdx;       // Top stone of stack
+    uint8_t squareIdx;         // Square
+    uint8_t topStoneIdx;       // Top stone of stack
     StoneType stoneType;   // F, S, C
     uint8_t stoneCount;    // Max: board width
     bool flattened;        // Capstone flattened standing
