+++ src/core/entities/History.h
@@ -3,6 +3,7 @@
 
 #include "GameConstants.h"
 #include "PlayerAction.h"
+#include <stdint.h>
 
 /**
  * @brief: Stack to store player turns
@@ -12,9 +13,9 @@ typedef struct
     /// Stack to store the actions to be undone
     PlayerAction actions[HISTORY_SIZE];
     /// Track index of last action that can be undone
-    int lastActionIdx;
+    uint8_t lastActionIdx;
     /// Track count of actions that can be redone
-    int redoCount;
+    uint8_t redoCount;
 } History;
 
 #endif
